Create FUNDING(waiting for a vote to see)
new to coding but trying to do this without knowledge sucks and i am learning. I want to create a blockchain community built off of a simple code langauge for people that cant code yet and make it easy for them to code. Also a fair community, and to build this together with anyone who wants in.